proper venv setup, audio-video shift repaired. refactored names
