+++ .idea/misc.xml
@@ -3,5 +3,5 @@
   <component name="Black">
     <option name="sdkName" value="Python 3.10 (et_procedure)" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.10 (et_procedure) (2)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.10 (et_procedure) (3)" project-jdk-type="Python SDK" />
 </project>
