+++ config.py
@@ -1,0 +1,14 @@
+CALIB_ANI_1_PATH = 'C://movies_et//norm_kalib_1.mp4'
+CALIB_ANI_2_PATH = 'C://movies_et//norm_kalib_2.mp4'
+CALIB_ANI_3_PATH = 'C://movies_et//norm_kalib_3.mp4'
+
+MOVIE_1_PATH = 'C://movies_et//norm_mov1.mp4'
+MOVIE_2_PATH = 'C://movies_et//norm_mov2.mp4'
+MOVIE_3_PATH = 'C://movies_et//norm_mov3.mp4'
+
+
+WIN_ID_MASTER = 1
+WIN_ID_MAIN = 0
+WIN_SIZES = [None, None]
+WIN_SIZES[WIN_ID_MASTER] = (1920, 1080)
+WIN_SIZES[WIN_ID_MAIN] = (2560, 1440)
