+++ m02_psychopy_routines.py
@@ -2,7 +2,7 @@ from psychopy.constants import (NOT_STARTED, STARTED, PLAYING, PAUSED,
                                 STOPPED, FINISHED, PRESSED, RELEASED, FOREVER)
 from psychopy import visual, core, event, sound
 import msgpack as serializer
-import comms
+import m03_pupilcapture_comms
 
 """
 These functions are mostly based on common code segments used in PsychoPy
